Module encoding_index_korean::euc_kr

KS X 1001 plus Unified Hangul Code.

From the Encoding Standard:

This matches the KS X 1001 standard and the Unified Hangul Code, more commonly known together as Windows Codepage 949.

Functions

backward
Returns the index pointer for code point code in this index.

forward
Returns the index code point for pointer code in this index.